+++ ServerDevcommands/Features/CommandLogging.cs
@@ -18,14 +18,15 @@ public class CommandLogging {
   public static string RPC_Log = "DEV_Log";
 
   private static string Format(ZNetPeer peer, string command) {
-    return String.Format(
+    return string.Format(
       Settings.Format(Settings.CommandLogFormat).Replace("command", "6"),
       peer.m_socket.GetHostName(), peer.m_playerName, peer.m_characterID.UserID.ToString(), peer.m_refPos.x, peer.m_refPos.y, peer.m_refPos.z, command
     );
   }
   private static void RPC_LogCommand(ZRpc rpc, string command) {
     var message = Format(ZNet.instance.GetPeer(rpc), command);
-    ZLog.Log(message);
+    if (message != "")
+      ZLog.Log(message);
   }
   static void Postfix(ZNet __instance, ZRpc rpc) {
     if (__instance.IsServer()) {